Split CAPI into sub-modules
