+++ modules/utils/index.js
@@ -1,15 +1,15 @@
 const env = require('./env-helper');
-const processSystem = require('./processSystem');
-const processBody = require('./processBody');
-const haversine = require('./haversine');
-const findRegion = require('./findRegion');
-const fetchRetry = require('./fetchRetry');
+// const processSystem = require('./processSystem');
+// const processBody = require('./processBody');
+// const haversine = require('./haversine');
+// const findRegion = require('./findRegion');
+// const fetchRetry = require('./fetchRetry');
 
 module.exports = {
 	env,
-	processSystem,
-	processBody,
-	haversine,
-	findRegion,
-	fetchRetry,
+	// processSystem,
+	// processBody,
+	// haversine,
+	// findRegion,
+	// fetchRetry,
 };
