+++ package.json
@@ -23,9 +23,7 @@
   "dependencies": {
     "axios": "^0.19.2",
     "cross-fetch": "^3.0.6",
-    "dotenv": "^8.2.0",
     "fetch-retry": "^2.2.3",
-    "find-config": "^1.0.0",
     "hookcord": "^2.0.2",
     "json-bigint": "^0.3.1",
     "lodash": "^4.17.20",
